chore(lint): security lint as a gate — ruff S + RUF, honest noqa, pre-commit
Enable ruff's bandit port (S) and the RUF family, with per-file-ignores
for what tests/, scripts/, alembic/ and seeders/ legitimately do, and
RUF001-003 off for the bilingual prose. Burn the findings down instead
of suppressing them: the ten mypy-narrowing asserts in src/ become
explicit raises, two fire-and-forget create_task calls hold a
reference (RUF006), one class-level mutable is a ClassVar, redundant
int(round()) casts go, __all__ lists are sorted. RUF100 removes the
130 noqa comments that silenced rules never enabled.

Bandit leaves the dev group — it was installed but ran nowhere; the
security-auditor agent now audits the S suppressions with
--ignore-noqa instead of re-discovering what CI gates. Add a
.pre-commit-config.yaml of local uv-run hooks (ruff on commit, mypy and
import-linter on push) so the pinned tools run before CI does.

Plan: docs/plans/lint-hardening.md (in-progress). uv.lock still needs
regenerating and mypy src/ re-running where the dev environment is
installed.

+++ src/kaleta/services/subscription_service.py
@@ -625,7 +625,7 @@ class SubscriptionService:
 def _amount_bucket(amount: Decimal) -> str:
     """Rough bucket for grouping amounts. Tight buckets → merged later."""
     # Round to the nearest 1 PLN; the ±5 % tolerance absorbs fine-grain drift.
-    return str(int(round(float(amount))))
+    return str(round(float(amount)))
 
 
 # Strip a trailing location/detail suffix so rows for the same service cluster.
